update deploy script

+++ deploy-to-npm.sh
@@ -4,13 +4,13 @@ set -euo pipefail
 PACKAGE_PATH="./dist/ez-vid-ang"
 
 # Ensure tag argument is provided
-#if [ $# -eq 0 ]; then
-#  echo "❌ Usage: ./release.sh <tag>"
-#  echo "Example: ./release.sh v19"
-#  exit 1
-#fi
+if [ $# -eq 0 ]; then
+  echo "❌ Usage: ./deploy-to-npm.sh <tag>"
+  echo "Example: ./deploy-to-npm.sh angular21"
+  exit 1
+fi
 
-#TAG="$1"
+TAG="$1"
 
 echo "🧹 Cleaning..."
 rm -rf .angular node_modules dist
@@ -28,6 +28,6 @@ echo "📦 Moving README.md"
 cp README.md dist/ez-vid-ang/README.md && cp LICENSE dist/ez-vid-ang/LICENSE
 
 echo "📤 Publishing to npm"
-npm publish "$PACKAGE_PATH"
+npm publish "$PACKAGE_PATH" --tag "$TAG"
 
 echo "🎉 Release successful"
